requirements.txt
streamlit
langchain
groq
pypdf2
docx2txt 
python-dotenv
langchain-groq